// src/app/models/defensive-game-stats.ts
export class DefensiveGameStats {
    gameId: number;
    playerId: number;
    tackles: number;
    tacklesForLoss: number;
    sacks: number;
    forcedFumbles: number;
    fumblesRecovered: number;
    interceptions: number;
    intYards: number;
    passesDeflected: number;
    touchdowns: number;
    safeties: number;
}